Enable console reuse

+++ vars/withConsole.groovy
@@ -9,19 +9,35 @@ def call(Map args, Closure body) {
     String consoleName = args.get('consoleName')
     String testbedName = args.get('testbed')
     String sonarProject = args.get('sonarProject', "master")
+    boolean useExistingConsole = args.get('useExistingConsole', false)
+
     sonarProject = sonarProject.isEmpty() ? "master" : sonarProject
 
     stage('Deploy console') {
 
-        withEnv(["AUTO_PARAMETERS=managementvm.console_name=${consoleName},${testbedName}"]) {
-            sh """#!/bin/bash
+        if (!useExistingConsole) {
+            withEnv(["AUTO_PARAMETERS=managementvm.console_name=${consoleName},${testbedName}"]) {
+                sh """#!/bin/bash
             source /auto/bin/env-setup.sh
             ManagementVm.py delete_all
             ManagementVm.py deploy
             """
+            }
+        } else {
+            withEnv(["AUTO_PARAMETERS=${testbedName}"]) {
+                def consoleNames = sh(returnStdout: true, script: '''#!/bin/bash
+                    source /auto/bin/env-setup.sh
+                    ManagementVm.py find
+            ''').trim()
+                echo "Found consoles: $consoleNames"
+                consoleName = (consoleNames =~ /[^\(]*\(([^\)]*).*/)[0][1]
+            }
+            echo "Using existing console: ${consoleName}"
         }
 
+
         if (collectCodeCoverage) {
+            echo "Configuring for code coverage data collection"
             withEnv(["AUTO_PARAMETERS=${testbedName},coverage.enable_console=True"]) {
                 sh '''#!/bin/bash
                 source /auto/bin/env-setup.sh
@@ -42,6 +58,7 @@ def call(Map args, Closure body) {
                 source /auto/bin/env-setup.sh
                 Triage.py copy_console_coverage'''
             }
+            echo "Code coverage data collected"
         } else {
             echo "Skipping code coverage collection; no data was generated"
         }
